refactor: remove obsolete Chromium tag group and well classes, consolidate functionality
